Add a sentencepiece tokenizer and make it default
This matches the tokenizer which was used to train LaMDA, according to the paper. This repo's tokenizer was trained on a small subset of The Pile's English Wikipedia dataset, totaling roughly 500k lines after lines that were too long for the tokenizer were discarded.
Testing with one run (which I will proceed to generalize to every run) shows that the loss is unaffected by the tokenizer, reduces the time needed to build the dataloader by half, and as a result of reducing the vocab size, reduces GPU memory usage. Very promising stuff.
Run: https://wandb.ai/waifu-collective-ai/lamda-hackernews-trainingtest/runs/30mfedeb
Next up: several experiments with scaling up, then experiments with other stuff.